secai-radar/api/requirements.txt
azure-functions>=1.18.0
azure-functions-durable>=1.2.8
azure-data-tables>=12.4.0
azure-storage-blob>=12.19.0
azure-keyvault-secrets>=4.7.0
azure-identity>=1.15.0
pydantic>=2.0.0
fastjsonschema>=2.19.0
python-multipart>=0.0.6
httpx>=0.27.0
openai>=1.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
# Multi-agent orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
# RAG Layer - Google File Search option
google-generativeai>=0.8.0
# Visualization
plotly>=5.18.0
# State persistence
azure-cosmos>=4.5.0
